+++ R/splice.R
@@ -5,8 +5,9 @@
 #'
 #' This splices all arguments into a list. Non-list objects and lists
 #' with a S3 class are encapsulated in a list before concatenation.
+#'
 #' We no longer believe that implicit/automatic splicing is a good idea;
-#' instead use `!!!` in conjunction with `rlang::list2()`.
+#' instead use `rlang::list2()` + `!!!` or [list_flatten()].
 #'
 #' @param ... Objects to concatenate.
 #' @return A list.
@@ -20,7 +21,7 @@
 #' c(inputs, arg3 = c("c1", "c2")) %>% str()
 #' @export
 splice <- function(...) {
-  lifecycle::deprecate_warn("0.4.0", "splice()")
+  lifecycle::deprecate_warn("0.4.0", "splice()", "list_flatten()")
 
   splice_if(list(...), is_bare_list)
 }
@@ -28,11 +29,5 @@ splice <- function(...) {
 splice_if <- function(.x, .p) {
   unspliced <- !probe(.x, .p)
   out <- modify_if(.x, unspliced, list)
-
-  # Copy outer names to inner
-  if (!is.null(names(.x))) {
-    out[unspliced] <- map2(out[unspliced], names(out)[unspliced], set_names)
-  }
-
-  flatten(out)
+  list_flatten(out, name_spec = "{inner}")
 }