+++ NEWS.md
@@ -18,7 +18,8 @@
   manipulation that is very uncommon in R code (#871).
 
 * `splice()` is deprecated because we no longer believe that automatic 
-  splicing makes for good UI. Instead use `list2()` + `!!!` (#869).
+  splicing makes for good UI. Instead use `list2()` + `!!!` or
+  `list_flatten()` (#869).
 
 * `as_function()`, `at_depth()`, and the `...f` argument to `partial()` 
   are no longer supported. They have been defunct for quite some time.
@@ -36,7 +37,16 @@
 * `*_raw()` have been deprecated because they are of limited use and you can 
   now use `map_vec()` instead (#903).
 
+* `flatten()` and friends are all deprecated in favour of `list_flatten()`, 
+  `list_c()`, `list_cbind()`, and `list_rbind()`.
+
+* `*_dfc()` and `*_dfr()` have been deprecated in favour of using the 
+  appropriate map function along with `list_rbind()` or `list_cbind()` (#912).
+
 ## Features and fixes
+
+* New `list_c()`, `list_rbind()`, and `list_cbind()` make it easy to
+  `c()`, `rbind()`, or `cbind()` all of the elements in a list.
 
 * `_lgl()`, `_int()`, `_int()`, and `_dbl()` now use the same (strict) coercion
   methods as vctrs (#904). This means that: